Change Styles of the Marker
