+++ packages/roosterjs-editor-dom/lib/list/VListItem.ts
@@ -1,16 +1,14 @@
 import contains from '../utils/contains';
-import ContentTraverser from '../contentTraverser/ContentTraverser';
 import getListTypeFromNode from './getListTypeFromNode';
-import getStyles from '../style/getStyles';
 import getTagOfNode from '../utils/getTagOfNode';
 import isBlockElement from '../utils/isBlockElement';
 import moveChildNodes from '../utils/moveChildNodes';
 import safeInstanceOf from '../utils/safeInstanceOf';
-import setStyles from '../style/setStyles';
 import toArray from '../utils/toArray';
 import unwrap from '../utils/unwrap';
 import wrap from '../utils/wrap';
-import { InlineElement, KnownCreateElementDataIndex, ListType } from 'roosterjs-editor-types';
+import { KnownCreateElementDataIndex, ListType } from 'roosterjs-editor-types';
+import { setListItemStyle } from '../utils/setListItemStyle';
 
 const orderListStyles = [null, 'lower-alpha', 'lower-roman'];
 
@@ -220,8 +218,12 @@ export default class VListItem {
 
         // 4. Inherit styles of the childelement to the li, so we are able to apply the styles to the ::marker
         if (this.listTypes.length > 1) {
-            const stylesToInherit = ['font-size', 'font-family', 'color'];
-            setListItemStyle(this.node, stylesToInherit);
+            if (
+                !(this.node.style.fontSize || this.node.style.color || this.node.style.fontFamily)
+            ) {
+                const stylesToInherit = ['font-size', 'font-family', 'color'];
+                setListItemStyle(this.node, stylesToInherit);
+            }
         }
 
         // 5. If this is not a list item now, need to unwrap the LI node and do proper handling
@@ -287,53 +289,5 @@ function wrapIfNotBlockNode(nodes: Node[], checkFirst: boolean, checkLast: boole
 function getChildrenAndUnwrap(node: Node): Node[] {
     const result = toArray(node.childNodes);
     unwrap(node);
-    return result;
-}
-
-function setListItemStyle(element: HTMLLIElement, stylesToInherit: string[]) {
-    if (element.style.fontSize || element.style.color || element.style.fontFamily) {
-        return;
-    }
-    const elementsStyles = getInlineChildElementsStyle(element);
-    let stylesToApply: Record<string, string> = getStyles(element);
-
-    stylesToInherit.forEach(styleName => {
-        const styleValues = elementsStyles.map(style =>
-            style[styleName] !== undefined ? style[styleName] : ''
-        );
-
-        if (
-            styleValues &&
-            (styleValues.length == 1 || new Set(styleValues).size == 1) &&
-            styleValues[0]
-        ) {
-            stylesToApply[styleName] = styleValues[0];
-        }
-    });
-    setStyles(element, stylesToApply);
-}
-function getInlineChildElementsStyle(element: HTMLElement) {
-    const result: Record<string, string>[] = [];
-    const contentTraverser = ContentTraverser.createBodyTraverser(element);
-    let currentInlineElement: InlineElement;
-
-    while (contentTraverser.currentInlineElement != currentInlineElement) {
-        currentInlineElement = contentTraverser.currentInlineElement;
-        let currentNode = currentInlineElement.getContainerNode();
-
-        if (currentNode.nodeName != 'BR') {
-            if (currentNode.nodeName == '#text') {
-                currentNode = currentNode.parentNode;
-            }
-            if (safeInstanceOf(currentNode, 'HTMLElement')) {
-                let childStyle = getStyles(currentNode);
-                if (childStyle) {
-                    result.push(childStyle);
-                }
-            }
-        }
-        contentTraverser.getNextInlineElement();
-    }
-
     return result;
 }
